add new generic transformer pipeline node
Specify any transformer sequence in a config file.
This includes any sklearn transformer via the SklearnPreprocessingTransformer
It's parent class is TestTrainSplit so it also retains those capabilities.

For this to work cleanly, I had to modify the test_train_split so that training fraction and seed are assigned as instance vars in the constructor and referenced as instance vars in the rest of the code instead of repeatedly being pulled from the node_config. Also it now pulls these params initially using dict get methods so the child class TransformerPipeline doesn't require those params in the config if you just want to use some transformers outside of a scenario where you need a test/train split. Also added a deprecation warning for the sklearn preprocessing pipeline since it's use cases are a subset of the new TransformerPipeline.